fixes a few more conversions that GCC10 on ubuntu22 had issues with - also catches a few more uses of `abs` rather than `std::abs`.

+++ mcmc/MinuitFit.cpp
@@ -37,7 +37,7 @@ void MinuitFit::runMCMC() {
   //KS: add config or something
   minuit->SetPrintLevel(2);
   minuit->SetTolerance(0.01);
-  minuit->SetMaxFunctionCalls(fitMan->raw()["General"]["Minuit2"]["NSteps"].as<double>());
+  minuit->SetMaxFunctionCalls(fitMan->raw()["General"]["Minuit2"]["NSteps"].as<unsigned>());
   minuit->SetMaxIterations(10000);
 
   MACH3LOG_INFO("Preparing Minuit");
